Added PWM control support via. Adafruit PCA9685
Added command parsing for easy control of the driver.
Added duty cycle control assuming that when duty cycle is less than half we are in reverse.
Added support for multiple servos using a dictionary to assign channel numbers to servo objects
Added readme to detail communication protocols

Added mocking of hardware to allow for services to be debugged on normal computers.

Added a script to test the I2C service